+++ src/BUILD.bazel
@@ -1,4 +1,4 @@
-load("@rules_python//python:defs.bzl", "py_binary")
+load("@rules_python//python:defs.bzl", "py_binary", "py_library")
 # load("@io_bazel_rules_docker//python3:image.bzl", "py3_image")
 
 py_binary(
@@ -6,6 +6,14 @@ py_binary(
     srcs = ["__main__.py"],
     imports = [".."],
     main = "__main__.py",
+    visibility = ["//:__subpackages__"],
+    deps = [":src"],
+)
+
+py_library(
+    name = "src",
+    srcs = ["__init__.py"],
+    imports = [".."],
     visibility = ["//:__subpackages__"],
     deps = ["@pip//pytermgui"],
 )
